requirements.txt
# requirements.txt
fastapi
uvicorn
kafka-python
aiokafka
networkx
osmnx
psycopg2-binary
APScheduler
requests
Flask
gunicorn
Flask-CORS
prometheus-client
psutil